# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QAA'A API",
        "version": "1.0.0",
        "description": "API for managing hall bookings and reservations"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token"
            }
        }
    },
    "paths": {
...
        "/api/bookings/{id}/confirm": {
            "post": {
                "tags": ["Bookings"],
                "summary": "Confirm booking payment",
                "operationId": "confirmPayment",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed"
                    }
                }
            }
        },
        "/api/owner/block-date": {
            "post": {
                "tags": ["Bookings"],
                "summary": "Block date for hall",
                "operationId": "blockDate",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "hall_id",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "hall_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Date blocked successfully"
                    }
                }
            }
        },
        "/api/my-bookings": {
            "get": {
                "tags": ["Bookings"],
                "summary": "Get customer's bookings",
                "operationId": "customerBookings",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of customer bookings"
                    }
                }
            }
        },
        "/api/owner/bookings": {
            "get": {
                "tags": ["Bookings"],
                "summary": "Get owner's bookings",
                "operationId": "ownerBookings",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of owner bookings"
                    }
                }
            }
        }
    }
}